Fix edge-case where CONCOURSE_TAG is empty on EC2 workers (#15)
We added the environment variable CONCOURSE_TAG on EC2 workers to
register those workers to only fetch jobs with those tags. The problem
comes when we don't need to specify any tag, then that variable is set
to empty, and somehow the ATC doesn't assign any job to that worker
because of that.

With this change, the worker tags are set as command line flags, and
multiple of them can be specified in a list.