Make the cross-account worker access more dynamic (#24)
* Make the cross-account worker access more dynamic

Previously the role name that the worker had to assume to access
the concourse keys on another account was hardcoded. That made it
hard to integrate in standard setups. With this change we provide
the full role ARN via a variable to make that integration more flexible.

* Update documentation for ec2-worker module

* Also create the IAM policy to allow the worker to assume the role for
cross account access

* Create the IAM role for cross-account access in the keys module