Configure the concourse vault role to issue periodic tokens
Currently, concourse can't use the AWS auth method to retrieve a Vault 
token, so we retrieve the token before starting the concourse web 
process via the Vault cli, and then provide that token to the concourse 
process. Because of that Concourse can't get a new token when this one 
expires, and the whole process has to be restarted. To avoid this, we've 
configured the Vault auth role to issue periodic tokens, which never 
expire as long as they are renewed in time (which concourse already 
does).

As per https://github.com/skyscrapers/engineering/issues/104